+++ modules/main.ts
@@ -9,6 +9,7 @@ import RC from './rc';
 import Starter from './starter';
 import Tester from './tester';
 import CFS from './cfs';
+import Panel from './panel';
 
 export default [
   Initializer,
@@ -22,4 +23,5 @@ export default [
   Starter,
   Tester,
   CFS,
+  Panel,
 ];