+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.9.0
+
+- Feat: added panel module.
+
+- Fix: `downloadFileWithProgress` method will stop progress bar first when an error ocurred.
+
 ## 0.8.0
 
 - Feat: added migrater module.